requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-cors-headers==3.6.0
django-extensions==3.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.2
Naked==0.1.31
packaging==20.8
pycrypto==2.6.1
pycryptodome==3.9.9
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
shellescape==3.8.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
python-dotenv
